Refactoring: Vytvorná AppPage.
